Create new Signer class and refactor
We create a new base class called Signer and make the HSM code and the
validation code a subclass of it.  This makes it a bit more obvious how
to implement support for additional types of signing infrastructure.

We also make a mockery of it for testing and extend the tests to use
said mockery.