Make a base class ChainRatchet and refactor
In order to make the code more modular, we define a ChainRatchet which
keeps track of the level of the last signed request.  If we are asked
to sign a request at <= that level, then we complain.

We then s/DynamoDBClient/DDBChainRatchet/.

We also make a mockery of ChainRatchet in the test program.