SignatureRequest no longer throws invalid message tag
Because of our restructure in 90f0af61f, it no longer makes sense
for SignatureRequest to throw an exception if it doesn't understand
the message tag.  We modify the code to parse as much of the sigreq
as possible and let ValidateSigner apply the policy of restricting
to only certain kinds of signatures.  In the future, we will have
to distinguish between sigreqs that require and don't require the
ratchet...